+++ src/map/camera.rs
@@ -70,7 +70,7 @@ impl Camera for Transform
     fn zoom(&mut self, units: f32)
     {
         let prev_scale = self.change_scale(units);
-        self.translate(-ui_camera_displacement() * (self.scale() - prev_scale));
+        self.translate(-ui_camera_displacement(self.scale() - prev_scale));
     }
 
     #[inline]
@@ -86,7 +86,7 @@ impl Camera for Transform
             padding
         );
 
-        self.set_pos(hull.center() - ui_camera_displacement() * self.scale());
+        self.set_pos(hull.center() - ui_camera_displacement(self.scale()));
     }
 }
 
@@ -134,5 +134,5 @@ pub fn scale_viewport(camera: &mut Transform, window_sizes: (f32, f32), hull: &H
 #[must_use]
 pub(in crate::map) fn init_camera_transform() -> Transform
 {
-    Transform::from_translation((-ui_camera_displacement()).extend(0f32))
+    Transform::from_translation((-ui_camera_displacement(1f32)).extend(0f32))
 }